refactor(common): share root space traversal utility
Extract root-space ancestry traversal into a reusable helper and consume it in
both the top menu branding and AI panel grouping flows to keep behavior aligned.

Made-with: Cursor
