+++ apps/web/src/components/connected-menu-top.tsx
@@ -5,7 +5,10 @@ import { useMemo } from 'react';
 import { usePathname } from 'next/navigation';
 import { useTranslations } from 'next-intl';
 import { MenuTop } from '@hypha-platform/ui';
-import { getDhoSpaceSlugFromPathname } from '@hypha-platform/epics';
+import {
+  getDhoSpaceSlugFromPathname,
+  getRootSpace,
+} from '@hypha-platform/epics';
 import useSWR from 'swr';
 import { Sparkles } from 'lucide-react';
 import { Space } from '@hypha-platform/core/client';
@@ -21,21 +24,6 @@ type ConnectedMenuTopProps = {
   closeMenuLabel?: string;
   aiChatEnabled: boolean;
 };
-
-function getRootSpace(activeSpace?: Space, spaces: Space[] = []) {
-  if (!activeSpace) return null;
-  const spacesById = new Map(spaces.map((space) => [space.id, space]));
-  let cursor: Space | undefined = activeSpace;
-  const seen = new Set<number>();
-  while (cursor?.parentId != null) {
-    if (seen.has(cursor.id)) break;
-    seen.add(cursor.id);
-    const parent = spacesById.get(cursor.parentId);
-    if (!parent) break;
-    cursor = parent;
-  }
-  return cursor ?? activeSpace;
-}
 
 function hasCustomRootLogo(logoUrl: string): boolean {
   return logoUrl.trim().length > 0;
